refactor booking-confirmation - split page into smaller components, hooks and utils
