+++ package.json
@@ -10,6 +10,7 @@
     "@testing-library/react": "^16.3.0",
     "@testing-library/user-event": "^14.6.1",
     "@types/express-session": "^1.18.2",
+    "@types/qrcode": "^1.5.6",
     "@types/react": "^19.2.2",
     "@types/react-bootstrap": "^1.0.1",
     "@types/react-dom": "^19.2.1",
@@ -29,6 +30,7 @@
     "jspdf": "^3.0.3",
     "mysql2": "^3.15.1",
     "nodemailer": "^7.0.10",
+    "qrcode": "^1.5.4",
     "react": "^19.2.0",
     "react-bootstrap": "^2.10.10",
     "react-bootstrap-icons": "^1.11.6",
