+++ src/pages/PagesStyle/BookingConfirmation.scss
@@ -1,4 +1,4 @@
-@use "./variables" as *;
+@use "../../styles/variables" as *;
 
 // Booking Confirmation Styling
